refactor: Adjust for Behaviors#89
A few components and systems have changed due to the switch from Pathfinding to FlexiblePathfinding.

This PR adjusts the breaking pieces of code to the best of my knowledge. Typical changes for this migration are:

- `MinionMoveComponent#target` is now a `Vector3i` and no longer a `Vector3f` (fixed by rounding)
- `MinionMoveComponent#type` was removed (no longer required, just remove references)

+++ src/main/java/org/terasology/taskSystem/actions/FindNextBlockOnBoundaryNode.java
@@ -3,6 +3,7 @@
 package org.terasology.taskSystem.actions;
 
 import org.joml.Vector3f;
+import org.joml.Vector3i;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.buildings.components.ConstructedBuildingComponent;
@@ -80,7 +81,7 @@ public class FindNextBlockOnBoundaryNode extends BaseAction {
         int y = taskComponent.taskRegion.minY();
 
         if (moveComponent.target == null) {
-            moveComponent.target = new Vector3f();
+            moveComponent.target = new Vector3i();
         }
         if (Math.round(worldPosition.x) == maxX) {
             if (Math.round(worldPosition.z) != maxZ) {
